+++ pkg/webhook/clusterresourceplacement/v1beta1_clusterresourceplacement_validating_webhook.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package clusterresourceplacement
 
 import (
 	"context"
